requirements.txt
tensorflow>=2.9.0
torch>=1.13.0
torchvision>=0.14.0
scikit-learn>=1.0.2
matplotlib>=3.5.1
pandas>=1.4.1
numpy>=1.22.2
opencv-python>=4.5.5
albumentations>=1.1.0
pillow>=9.0.1
jupyter>=1.0.0
tqdm>=4.62.3
seaborn>=0.11.2
